refactor: Change rating data type from integer to Double

+++ app/src/main/java/org/example/factory/DishReviewFactory.java
@@ -7,12 +7,12 @@ import org.example.models.reviews.Review;
 public class DishReviewFactory implements IReviewFactory<Dish>{
 
   @Override
-  public Review createReview(int rating, String comment, Dish dish, Integer... additionalRatings) {
+  public Review createReview(Double rating, String comment, Dish dish, Double... additionalRatings) {
     if(additionalRatings.length < 2){
       throw new IllegalArgumentException("Faltan calificaciones adicionales");
     }
-    Integer flavorRating = additionalRatings[0];
-    Integer presentationRating = additionalRatings[1];
+    Double flavorRating = additionalRatings[0];
+    Double presentationRating = additionalRatings[1];
     return new DishReview(rating, comment, dish,flavorRating,presentationRating);
   }
 }
