+++ app/src/main/java/org/example/utils/NotificationService.java
@@ -29,7 +29,7 @@ public class NotificationService implements IObservable {
       observer.update(message);
     }
   }
-  public void notifyNewReview(String entityName, String entityType, Integer rating){
+  public void notifyNewReview(String entityName, String entityType, Double rating){
     String message = "Nueva review añadida para " + entityType + ": " + entityName + ". con una calificación: " + rating;
     notifyObservers(message);
   }
